Initial Commit
This git repository is being created from a project that was already in-
progress, and thus has previous work undocommented by commit. Additionally,
this repository will be pushed to github shortly after this initial commit.
Currently, the project is NOT in a usable state, and has a current list of
goals documented in the README.md.