It compiles!
Added in a couple more source files from the support library that
were needed. Seems to be a couple warnings, one telling me about
the timing (not sure why this is a warning), and one telling me
about some UART related error, but this is not (yet?) using UART,
so I'm not too concerned about that atm.