Driver correctly controls i2c i/o expander
So far the driver can correctly address the individual pins on the
i2c i/o expander that my lcd uses, and the driver appears to be
setting the pins correctly to control it, but the lcd still isn't
working. Not sure whether or not the lcd is any good. May see if
I can find the extra hardware and an already existing test program
for a RPi to see if I need to buy a new lcd or if I'm doing
something wrong still.